MG was always truncating but it is better to test exactly when needed

+++ sarra/sr_util.py
@@ -160,7 +160,11 @@ class sr_proto():
         # flush,sync, remember current position, truncate = no sparce, close
 
         self.fpos = dst.tell()
-        dst.truncate()
+
+        if not hasattr(self.parent,'msg')   : dst.truncate()
+        elif self.parent.msg.lastchunk      : dst.truncate()
+        elif self.parent.msg.partflg == 'p' : dst.truncate()
+
         dst.close()
 
         # finalize checksum
